Modification de type dans create_database.sql
- modification du type de
  ProfileManagementData.profile_management_data_salary de float à int.

+++ Web/create_database.sql
@@ -66,7 +66,7 @@ UPDATE Profiles SET profile_password = MD5('genymobile') WHERE profile_id=5;
 CREATE TABLE ProfileManagementData (
 	profile_management_data_id int auto_increment,
 	profile_id int not null unique,
-	profile_management_data_salary float not null,
+	profile_management_data_salary int not null,
 	profile_management_data_recruitement_date date not null,
 	profile_management_data_is_billable boolean not null default true,
 	profile_management_data_availability_date date not null,
